refactor: split my/+layout.svelte and my/friends/+page.svelte under 200 lines
Extracts orchestration logic out of components per the project's 200-line
guideline:
- src/lib/state/boardSync.svelte.ts: refreshFromServer reconcile/retry loop,
  cache hydration, and the debounced local-cache persistence effects (moved
  out of my/+layout.svelte, now 131 lines, down from 208)
- src/lib/state/friendsActions.ts: the 4 optimistic friend/invite handlers
  (moved out of my/friends/+page.svelte)
- src/components/FriendListItem.svelte: the repeated friend/invite row
  markup, previously an inline {#snippet} (my/friends/+page.svelte is now
  172 lines, down from 263)

Fixes #777

No behavior changes - purely moving existing logic into dedicated modules
so it's usable/testable independent of component markup.

Verified manually via agent-browser: board hydration + server reconciliation
+ note edit/save still persist correctly across reload, and all 4 friends
handlers (cancel/remove/accept/reject) still work via the extracted
FriendListItem component, using seeded friend/invite test data.
